Extract use cases to separate module
