--- floppa/telegram/commands/custom/del_command/__init__.py
@@ -1,1 +1,0 @@
-import floppa.telegram.commands.custom.del_command.handler
