--- floppa/telegram/commands/custom/set_command/__init__.py
@@ -1,1 +1,0 @@
-import floppa.telegram.commands.custom.set_command.handler
