+++ floppa/use_case/__init__.py
@@ -1,0 +1,2 @@
+from floppa.use_case.del_command import DeleteCommandUseCase
+from floppa.use_case.set_command import SetCommandUseCase
