+++ floppa/use_case/del_command/__init__.py
@@ -1,0 +1,1 @@
+from floppa.use_case.del_command.del_command import DeleteCommandUseCase
