+++ floppa/use_case/set_command/__init__.py
@@ -1,0 +1,1 @@
+from floppa.use_case.set_command.set_command import SetCommandUseCase
